motion ref handler module
